+++ java/clients/src/main/java/sleeper/clients/status/report/compaction/job/CompactionJobStatusReporter.java
@@ -17,7 +17,7 @@
 package sleeper.clients.status.report.compaction.job;
 
 import sleeper.clients.status.report.job.query.JobQuery;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 
 import java.util.List;
