+++ java/clients/src/main/java/sleeper/clients/status/report/compaction/job/StandardCompactionJobStatusReporter.java
@@ -24,9 +24,9 @@ import sleeper.clients.util.table.TableFieldDefinition;
 import sleeper.clients.util.table.TableRow;
 import sleeper.clients.util.table.TableWriter;
 import sleeper.clients.util.table.TableWriterFactory;
-import sleeper.compaction.core.job.status.CompactionJobCommittedStatus;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
-import sleeper.compaction.core.job.status.CompactionJobStatusType;
+import sleeper.compaction.core.job.query.CompactionJobCommittedStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatusType;
 import sleeper.core.record.process.AverageRecordRate;
 import sleeper.core.record.process.status.ProcessRun;
 import sleeper.core.util.DurationStatistics;
