+++ java/clients/src/main/java/sleeper/clients/status/report/job/query/DetailedJobsQuery.java
@@ -16,7 +16,7 @@
 package sleeper.clients.status.report.job.query;
 
 import sleeper.compaction.core.job.CompactionJobStatusStore;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
