+++ java/clients/src/main/java/sleeper/clients/status/report/job/query/RejectedJobsQuery.java
@@ -17,7 +17,7 @@
 package sleeper.clients.status.report.job.query;
 
 import sleeper.compaction.core.job.CompactionJobStatusStore;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
