+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/query/CompactionJobStartedStatus.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.core.job.status;
+package sleeper.compaction.core.job.query;
 
 import sleeper.core.record.process.status.ProcessRunStartedUpdate;
