+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/job/CompactionJobStatusFurthestStatusTest.java
@@ -17,11 +17,11 @@ package sleeper.compaction.core.job;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.core.job.status.CompactionJobCommittedStatus;
-import sleeper.compaction.core.job.status.CompactionJobCreatedStatus;
-import sleeper.compaction.core.job.status.CompactionJobFinishedStatus;
-import sleeper.compaction.core.job.status.CompactionJobStartedStatus;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobCommittedStatus;
+import sleeper.compaction.core.job.query.CompactionJobCreatedStatus;
+import sleeper.compaction.core.job.query.CompactionJobFinishedStatus;
+import sleeper.compaction.core.job.query.CompactionJobStartedStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.core.record.process.status.ProcessFailedStatus;
 
 import java.time.Duration;
@@ -33,11 +33,11 @@ import static sleeper.compaction.core.job.CompactionJobStatusTestData.compaction
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.compactionFinishedStatus;
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.compactionStartedStatus;
 import static sleeper.compaction.core.job.CompactionJobStatusTestData.jobStatusFromUpdates;
-import static sleeper.compaction.core.job.status.CompactionJobStatusType.CREATED;
-import static sleeper.compaction.core.job.status.CompactionJobStatusType.FAILED;
-import static sleeper.compaction.core.job.status.CompactionJobStatusType.FINISHED;
-import static sleeper.compaction.core.job.status.CompactionJobStatusType.IN_PROGRESS;
-import static sleeper.compaction.core.job.status.CompactionJobStatusType.UNCOMMITTED;
+import static sleeper.compaction.core.job.query.CompactionJobStatusType.CREATED;
+import static sleeper.compaction.core.job.query.CompactionJobStatusType.FAILED;
+import static sleeper.compaction.core.job.query.CompactionJobStatusType.FINISHED;
+import static sleeper.compaction.core.job.query.CompactionJobStatusType.IN_PROGRESS;
+import static sleeper.compaction.core.job.query.CompactionJobStatusType.UNCOMMITTED;
 import static sleeper.core.record.process.RecordsProcessedSummaryTestHelper.summary;
 import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.failedStatus;
