+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/job/CompactionJobStatusTaskIdAssignedTest.java
@@ -17,7 +17,7 @@ package sleeper.compaction.core.job;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.core.record.process.status.ProcessRun;
 
 import java.time.Instant;
