+++ java/compaction/compaction-status-store/src/main/java/sleeper/compaction/status/store/job/DynamoDBCompactionJobStatusFormat.java
@@ -20,11 +20,11 @@ import org.apache.commons.codec.binary.Hex;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.compaction.core.job.status.CompactionJobCommittedStatus;
-import sleeper.compaction.core.job.status.CompactionJobCreatedStatus;
-import sleeper.compaction.core.job.status.CompactionJobFinishedStatus;
-import sleeper.compaction.core.job.status.CompactionJobStartedStatus;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobCommittedStatus;
+import sleeper.compaction.core.job.query.CompactionJobCreatedStatus;
+import sleeper.compaction.core.job.query.CompactionJobFinishedStatus;
+import sleeper.compaction.core.job.query.CompactionJobStartedStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.core.record.process.ProcessRunTime;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
