+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/reporting/CompactionReportsDriver.java
@@ -16,7 +16,7 @@
 
 package sleeper.systemtest.dsl.reporting;
 
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.compaction.core.task.CompactionTaskStatus;
 
 import java.util.List;
