+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/util/WaitForJobsStatus.java
@@ -22,8 +22,8 @@ import com.google.gson.JsonSerializer;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import sleeper.compaction.core.job.CompactionJobStatusStore;
-import sleeper.compaction.core.job.status.CompactionJobStatus;
-import sleeper.compaction.core.job.status.CompactionJobStatusType;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatusType;
 import sleeper.core.record.process.status.ProcessRun;
 import sleeper.core.util.GsonConfig;
 import sleeper.ingest.core.job.status.IngestJobStatus;
