+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemoryReports.java
@@ -15,7 +15,7 @@
  */
 package sleeper.systemtest.dsl.testutil.drivers;
 
-import sleeper.compaction.core.job.status.CompactionJobStatus;
+import sleeper.compaction.core.job.query.CompactionJobStatus;
 import sleeper.compaction.core.task.CompactionTaskStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
